Smarter k8s deployment.

+++ project/DeployPlugin.scala
@@ -9,17 +9,34 @@ object DeployKeys {
 
 object DeployPlugin extends AutoPlugin {
   override val projectSettings = Seq(
-    DeployKeys.deployLocal := {
-      val dockerImageID = DockerKeys.dockerBuildAndPush.value
+    DeployKeys.deployLocal := Def.taskDyn {
+      (sbt.Keys.compile in Compile).value // make sure the entire compile finishes first
 
-      // Remove the created image; we don't need it once it has been pushed to the registry
-      val rmi = Process("docker rmi " + dockerImageID).!!
-      println(rmi)
+      // Does this project have a mainClass we can run?
+      val hasMainClass = (mainClass in run in Compile).value.nonEmpty
 
-      // Delete pods associated with this deployment, so they spin back up with the new version
-      val delete = Process("kubectl delete pods -l service=" + baseDirectory.value.getName()).!!
-      println(delete)
-    }
+      // Did the incremental compile actually do anything?
+      val incrementalCompileDidSomething = (sbt.Keys.compileIncremental in Compile).value.hasModified
+
+      if (hasMainClass && incrementalCompileDidSomething) {
+        Def.task {
+          val dockerImageID = DockerKeys.dockerBuildAndPush.value
+
+          // Remove the created image; we don't need it once it has been pushed to the registry
+          val rmi = Process("docker rmi " + dockerImageID).!!
+          println(rmi)
+
+          // Delete pods associated with this deployment, so they spin back up with the new version
+          val delete = Process("kubectl delete pods -l service=" + baseDirectory.value.getName()).!!
+          println(delete)
+          true
+        }
+      } else {
+        Def.task {
+          false
+        }
+      }
+    }.value
   )
 
   override def requires = sbtdocker.DockerPlugin
